scripts/requirements.txt
# Python dependencies for paper automation system
# Install with: pip install -r requirements.txt

# Core dependencies
arxiv>=2.0.0
PyYAML>=6.0
requests>=2.31.0

# AI and ML
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.0.1

# Audio generation
edge-tts>=6.1.9

# Vector database (for Q&A)
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Data processing
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0

# Semantic Scholar API
semanticscholar>=0.4.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
